xen/arm: Restrict "p2m_ipa_bits" according to the IOMMU requirements
There is a strict requirement for the IOMMU which wants to share
the P2M table with the CPU. The IOMMU's Stage-2 input size must be equal
to the P2M IPA size. It is not a problem when the IOMMU can support
all values the CPU supports. In that case, the IOMMU driver would just
use any "p2m_ipa_bits" value as is. But, there are cases when not.

In order to make P2M sharing possible on the platforms which
IOMMUs have a limitation in maximum Stage-2 input size introduce
the following logic.

First initialize the IOMMU subsystem and gather requirements regarding
the maximum IPA bits supported by each IOMMU device to figure out
the minimum value among them. In the P2M code, take into the account
the IOMMU requirements and choose suitable "pa_range" according
to the restricted "p2m_ipa_bits".

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
CC: Julien Grall <julien.grall@arm.com>

---
Please note, this patch wasn't checked for the SMMU.

Changes V4 [4] -> V5:
   - [SMMU] Use p2m_ipa_bits instead of smmu->s2_input_size for
     the TTBCR calculation

Changes RFC V3 [3] -> V4 [4]:
   - Move check for p2m_ipa_bits to be at least 40 bit
     under #ifdef CONFIG_ARM_32
   - Reword the "panic" message

Changes RFC V2 [2] -> RFC V3 [3]:
   - Check in setup_virt_paging() that the "restricted"
     P2M IPA size is at least 40-bit
   - Modify logic in setup_virt_paging() a bit to make it
     "IOMMU-agnostic"
   - Clarify comments in code, add some explanations
   - Avoid using the term "IOMMU" in P2M code where possible

Changes RFC V1 [1] -> RFC V2 [2]:
   - Don't update p2m_ipa_bits by the IOMMU drivers directly,
     introduce p2m_restrict_ipa_bits()
   - Clarify patch subject/description
   - Add more comments to code
   - Check for equivalent "pabits" in setup_virt_paging()
   - Remove ASSERTs from the SMMU and IPMMU drivers

[1] https://lists.xenproject.org/archives/html/xen-devel/2019-08/msg02078.html
[2] https://lists.xenproject.org/archives/html/xen-devel/2019-08/msg02237.html
[3] https://lists.xenproject.org/archives/html/xen-devel/2019-09/msg00973.html
[4] https://lists.xenproject.org/archives/html/xen-devel/2019-09/msg02304.html
---