ioreq: Drop Arm restriction for buffered ioreq
The restriction is in place because it was decided to not let
the buffered I/O feature to be used on other than x86 arches
in Xen for now.

But Qemu setups buffered ioreq to be used on x86. It is fine
to drop the check here as it is not really used on Arm for
the communication at the runtime, it is just initialized and all.
But the proper solution would be to update the Qemu to not setup
buffered ioreq on Arm.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>