arm/vgic: provide late_init function
vGIC ITS wants to map ITS registers into a guest, but at the moment of
domain_vgic_init, p2m is not initialized yet, to so any call to p2m
will fail. So, we need to postpone ITS initialization. This patch
adds domain_late_init call-back, which can be called when domain
construction is complete.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>