[WA?]vpci: fix MMIO handler cound calculation for the hardware dom
Hardware domain can any number of PCI devices, so it is not feasible
to make correct estimate of number MMIO handlers needed for
it. This patch makes assumption that every possible device on a host
bridge an have MSI-X enabled, so we would report worst case to the
caller.

This is not upstreamable and we need to rework approach to MMIO
handlers. Maybe - allocate them dynamically or something...

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>